- Split app.py server.py setting.py - Fix `ping` API accept ends with '/' and '' to avoid forward to another system
