--- src/envs.py
@@ -1,4 +1,0 @@
-TOKEN_ENV = 'RUBIX_SERVICE_TOKEN'
-DATA_DIR_ENV = 'RUBIX_SERVICE_DATA'
-ARTIFACT_DIR_ENV = 'ARTIFACT_DIR'
-GLOBAL_DATA_DIR_ENV = 'GLOBAL_DATA'
